requirements.txt
# Core dependencies
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
httpx>=0.24.0

# Database dependencies
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# Utility dependencies
numpy>=1.24.0
python-dotenv>=1.0.0

# Moderation dependencies
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1  # For better performance with fuzzywuzzy

# Web scraping dependencies
beautifulsoup4>=4.9.0
requests>=2.25.0
lxml>=4.6.0
aiohttp>=3.8.0

# MCP dependencies
mcp>=1.6.0
websockets>=10.0

# Monitoring dependencies
langfuse>=2.0.0

# Tavily dependencies
tavily-python>=0.2.0

# Document processing dependencies
PyPDF2>=2.0.0
python-docx>=0.8.10
pdfplumber>=0.5.0
tqdm>=4.60.0
nltk>=3.6.0
